% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_utils.R
\name{divide}
\alias{divide}
\title{Divide two numbers}
\usage{
divide(x1, x2)
}
\arguments{
\item{x1}{Numerator}

\item{x2}{Denominator. Must not be zero.}
}
\value{
Numeric result of x1 / x2
}
\description{
Divide two numbers
}
