% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_utils.R
\name{percent_change}
\alias{percent_change}
\title{Calculate percent change between two values}
\usage{
percent_change(old_val, new_val)
}
\arguments{
\item{old_val}{Starting value. Must not be zero.}

\item{new_val}{Ending value.}
}
\value{
Numeric percent change from old_val to new_val.
}
\description{
Calculate percent change between two values
}
